F-4339: widen uart-flash-server bounds checks to 64-bit to stop uint32_t wrap
uart_flash_erase/read/write read attacker-controlled address and len as raw
4-byte words from the UART peer and guard the mmap region with
`address + len > FIRMWARE_PARTITION_SIZE + SWAP_SIZE`. The addition is done in
uint32_t, so a large len (e.g. 0xFFFFFFFF) wraps the sum below 0x21000 and the
guard passes, after which the loop walks far past the 0x21000-byte mapping —
SIGSEGV on a 64-bit host, or wrapped writes over the firmware header on 32-bit.

Compute address + len in uint64_t before the comparison so the sum cannot wrap.
Add a pty-driven regression test (test_overflow) and a `make test` target that
sends an ERASE with a wrapping address+len and asserts the server survives.

+++ tools/uart-flash-server/ufserver.c
@@ -280,7 +280,7 @@ static void uart_flash_erase(uint8_t *base, int ud)
         return;
     if (read_word(ud,&len) != 4)
         return;
-    if (address + len > (FIRMWARE_PARTITION_SIZE + SWAP_SIZE))
+    if ((uint64_t)address + (uint64_t)len > (FIRMWARE_PARTITION_SIZE + SWAP_SIZE))
         return;
     if (address < FIRMWARE_PARTITION_SIZE) {
         printmsg(msgEraseUpdate);
@@ -319,7 +319,7 @@ static void uart_flash_read(uint8_t *base, int ud)
 #if LOG_FLASH_ADDRESS
     printf("Read @%x\n", address);
 #endif
-    if (address + len > (FIRMWARE_PARTITION_SIZE + SWAP_SIZE))
+    if ((uint64_t)address + (uint64_t)len > (FIRMWARE_PARTITION_SIZE + SWAP_SIZE))
         return;
     for (i = 0; i < len; i++) {
         write(ud, base + address + i, 1);
@@ -344,7 +344,7 @@ static void uart_flash_write(uint8_t *base, int ud)
 #if LOG_FLASH_ADDRESS
     printf("Write @%x\n", address);
 #endif
-    if (address + len > (FIRMWARE_PARTITION_SIZE + SWAP_SIZE))
+    if ((uint64_t)address + (uint64_t)len > (FIRMWARE_PARTITION_SIZE + SWAP_SIZE))
         return;
     for (i = 0; i < len; i++) {
         read(ud, base + address + i, 1);
